requirements.txt
Django==3.1.1
celery==5.0.0
psycopg2==2.8.6
djangorestframework==3.12.1
django-split-settings==1.0.1
